Moves more user related repos to Shoebox Module
