+++ kifi-backend/modules/common/test/com/keepit/common/store/FakeStoreModule.scala
@@ -1,14 +1,12 @@
 package com.keepit.common.store
 
 import scala.collection.mutable.HashMap
-import scala.concurrent.{Future, promise}
 
-import com.amazonaws.services.s3.model.PutObjectResult
 import com.google.inject.{Singleton, Provides}
-import com.keepit.common.db.{ExternalId, Id}
+import com.keepit.common.db.Id
 import com.keepit.common.social.SocialUserRawInfo
 import com.keepit.common.social.SocialUserRawInfoStore
-import com.keepit.model.{User, NormalizedURI, SocialUserInfo}
+import com.keepit.model.{NormalizedURI, SocialUserInfo}
 import com.keepit.search.{InMemoryArticleSearchResultStoreImpl, ArticleSearchResultStore, Article, ArticleStore}
 import com.keepit.common.analytics.{FakeMongoS3EventStore, MongoEventStore}
 
@@ -21,19 +19,9 @@ trait FakeStoreModule extends StoreModule {
   }
 
   @Provides @Singleton
-  def s3ImageStore(s3ImageConfig: S3ImageConfig): S3ImageStore = FakeS3ImageStore(s3ImageConfig)
-
-  @Provides @Singleton
   def fakeMongoStore() : MongoEventStore = new FakeMongoS3EventStore()
 
   @Provides @Singleton
   def articleSearchResultStore(): ArticleSearchResultStore = new InMemoryArticleSearchResultStoreImpl()
 
-}
-
-case class FakeS3ImageStore(val config: S3ImageConfig) extends S3ImageStore {
-  def updatePicture(sui: SocialUserInfo, externalId: ExternalId[User]) =
-    promise[Seq[PutObjectResult]]().success(Seq()).future
-  def getPictureUrl(w: Int, user: User) =
-    promise[String]().success(s"http://cloudfront/${user.id.get}_${w}x${w}").future
 }
