+++ kifi-backend/modules/common/test/com/keepit/test/DeprecatedDbRepos.scala
@@ -9,16 +9,10 @@ import com.keepit.common.db.TestSlickSessionProvider
 trait DeprecatedDbRepos { self: InjectorProvider =>
 
   def db(implicit injector: Injector) = inject[Database]
-  def userSessionRepo(implicit injector: Injector) = inject[UserSessionRepo]
-  def userConnRepo(implicit injector: Injector) = inject[UserConnectionRepo]
-  def socialConnRepo(implicit injector: Injector) = inject[SocialConnectionRepo]
   def urlRepo(implicit injector: Injector) = inject[URLRepo]
   def socialUserInfoRepo(implicit injector: Injector) = inject[SocialUserInfoRepo]
   def installationRepo(implicit injector: Injector) = inject[KifiInstallationRepo]
-  def userExperimentRepo(implicit injector: Injector) = inject[UserExperimentRepo]
-  def emailAddressRepo(implicit injector: Injector) = inject[EmailAddressRepo]
   def unscrapableRepo(implicit injector: Injector) = inject[UnscrapableRepo]
-  def notificationRepo(implicit injector: Injector) = inject[UserNotificationRepo]
   def collectionRepo(implicit injector: Injector) = inject[CollectionRepo]
   def sessionProvider(implicit injector: Injector) = inject[SlickSessionProvider].asInstanceOf[TestSlickSessionProvider]
