+++ kifi-backend/modules/shoebox/app/com/keepit/common/store/S3ImageStore.scala
